Refactor out Spark 3 Migration
